Remove pytorch model dependency
This commit loads many pytorch dependent variables and methods from the
provided model object. This way, any model (not only torch) can be used
with CP as long as the model object provides the required elements.

This is the first step toward building our own model SDK for generic
usability and scalable support.